feat(*) add git-rev-like support for heroku with SOURCE_VERSION
When pushing to heroku using git to deploy, the .git folder is deleted before compiling the slug, so git-rev-parse fails.
Though, the hash of the revision is provided by heroku in the env var SOURCE_VERSION during the compiling time.
The hash is saved into bannerInfos.json to be available to express so that it could render the hash in the html (at this time the app will already launched and the env var not available anymore).

Yes :-) All this to get the revision in a comment that nobody will ever read ! LOL